Fixed the code to send sensor information up.
The code had a bug where I was casting the int of information into a byte,
bit-shifting it, and then sending up the information. Now I have fixed the
code so that it first does a bit shift, and then does an AND on it.

And yes, I do realize that I probably don't don't need that AND, but you
can change that yourself.