Added an additional motor id and a warning
Essentially, we had an indexing error. There were 8 motors in the array,
and I was calling the 8th motor. The highest index should've been 7. Now,
I have added in a warning to prevent errors like this in the future.

Also, changed some stepper stuff. This will be changed shortly, as testing
has revealed that using a PWM pin will give us the best results. Expect
some changes in the protocol!